Use token ranges

+++ UrlShortener.Api.Core.Tests/Base62EncodingScenarios.cs
@@ -1,4 +1,6 @@
 ﻿
+using UrlShortener.Core;
+
 namespace UrlShortener.Api.Core.Tests;
 
 public class Base62EncodingScenarios
@@ -10,7 +12,7 @@ public class Base62EncodingScenarios
     [InlineData(1000, "G8")]
     [InlineData(61, "z")]
     [InlineData(987654321, "14q60P")]
-    public void Should_Encode_Number_To_Base62(int number, string expected)
+    public void Should_Encode_Number_To_Base62(long number, string expected)
     {
         number.EncodeToBase62().Should().Be(expected);
     }
